Update math library function hooking to hook address instead of symbol name.
Previously, when we hook function calls, we hook an address, the call target.
For a library function, this is the address in the PLT.
When we hook the symbol of the math library function, the hook in the PLT will execute first, therefore short circuiting the hook we want.
Now, we hook the address in the PLT of the math functions and skip hooking call targets that have been hooked already.