Update function call short circuiting
Update short_circuit_calls to be a user supplied map from a function's address to its parameter and return types.
This map is used to replace calls to this function with a new claripy operation, named after the function.
The arguments passed to the function are passed as operands to this claripy operation.